Revert "sandbox: remove os_putc() and os_puts()"
While sandbox works OK without the special-case code, it does result in
console output being stored in the pre-console buffer while sandbox starts
up. If there is a crash or a problem then there is no indication of what
is going on.

For ease of debugging it seems better to revert this change.

This reverts commit 6f63f5449f420175c32e30ea176f7037eb2d8a6c.

Signed-off-by: Simon Glass <sjg@chromium.org>