Revert "sandbox: Drop special case console code for sandbox"
While sandbox works OK without the special-case code, it does result in
console output being stored in the pre-console buffer while sandbox starts
up. If there is a crash or a problem then there is no indication of what
is going on.

For ease of debugging it seems better to revert this change also.

This reverts commit dbd092656c9d3263c17a0f540f9fb7be7fdb8da3.

Signed-off-by: Simon Glass <sjg@chromium.org>